requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
executor==23.2
fasteners==0.19
frozenlist==1.4.1
humanfriendly==10.0
idna==3.7
libretranslatepy==2.1.1
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
property-manager==3.0
pydantic==2.7.3
pydantic_core==2.18.4
pyreadline3==3.4.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
translate==3.6.1
typing_extensions==4.12.1
urllib3==2.2.1
verboselogs==1.7
yarl==1.9.4
